Initial data commit
x3_y3_z1 means 3 gyroid unit cells in x & y directions, and 1 in z direction.
90 mm is the length of 3 unit cells
Mod means "modified" - this specimen had flaps added to the sides and a cylinder at the bottom to interface with the steel plate and shaker stinger, respectively.
018VF means 18% volume fraction of the gyroid (dependent on its thickness)
ABS is the material - Acrylonitrile butadiene styrene, a common FDM (type of 3D printing) material
20200706 is the date - 6th of July 2020

Files are sorted by specimen and subsequently by Run number. Note that not all runs had the same conditions in every folder in initial specimens - the experiment setup was still not finalised and frozen.